+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 ## develop
 
+## v5.0.0
+
+Released Wednesday, 30th October 2019.
+
+### Backwards-Compatibility Breaks
+
+* Dropped `Pipe.Vars`
+* `Pipe.Env` now defaults to the program's environment
+  - you can replace it with an `OverlayEnv` from Envish if you'd like
+* Dropped `Pipe.Getvar()`
+  - call `Pipe.Env.Getenv()` instead
+* Dropped `Pipe.Expand()`
+  - call `Pipe.Env.Expand()` instead
+
+### Dependencies
+
+- bumped to go_envish v3.0.1
+
 ## v4.3.0
 
 Released Tuesday, 8th October 2019.